use type system to enforce child index limit and key path depth limit
BIP32 key path indexes for unhardened (public) key derivation cannot
exceed 0x7FFFFFFF, so we add ChildIndex to represent numbers of this
class. Furthermore BIP32 keys cannot have a depth of more than 255
as they must be represented in serialized xpubs as a single byte,
so we introduce a KeyPath type to enforce that limit using the type
system.